requirements.txt
dateparser==1.1.8
requests==2.28.1
schedule==1.1.0
meross-iot==0.4.5.4
redis==3.5.3
paho-mqtt==1.6.1
beautifulsoup4==4.11.1
tinytuya==1.2.8
openai==0.27.2
boto3==1.26.14
botocore==1.29.24
RPi.GPIO==0.7.0
six==1.16.0
google-api-core==2.0.1
google-auth==2.14.1
google-cloud-speech==2.8.0
googleapis-common-protos==1.57.0
pvporcupine==2.1.4
PyAudio==0.2.12
